Dashboard

+++ app/src/main/java/com/example/kidsindata_spaceinvader/model/ScoringTrend.kt
@@ -8,5 +8,5 @@ data class ScoringTrend (
 
     @SerializedName("gameDateTime") var gameDateTime: String,
 
-    @SerializedName("playerScore") var playerScore: Int
+    @SerializedName("playerScore") var playerScore: Double
 )
